feat(batch-d): convert fcm-tokens, universities, banners, sales, app-reviews to v2
